+++ ft_memset.c
@@ -6,11 +6,13 @@
 /*   By: tlegendr <tlegendr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/17 16:36:07 by tlegendr          #+#    #+#             */
-/*   Updated: 2023/10/19 12:34:07 by tlegendr         ###   ########.fr       */
+/*   Updated: 2023/11/04 17:31:34 by tlegendr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *pointer, int value, unsigned int count)
+#include "libft.h"
+
+void	*ft_memset(void *pointer, int value, size_t count)
 {
 	char			*ptr;
 	unsigned int	i;
